+++ src/units.rs
@@ -339,13 +339,13 @@ impl<D: Direction> UnitType<D> {
             Self::Structure(structure) => structure.typ.value(),
         }
     }
-    pub fn value(&self, game: &Game<D>, _co: &Commander) -> u16 {
+    pub fn value(&self, game: &Game<D>, _co: &Commander) -> usize {
         (match self {
             Self::Normal(unit) => unit.typ.value(),
             Self::Mercenary(merc) => merc.unit.typ.value() + merc.typ.price(game, &merc.unit).unwrap() as u16,
             Self::Chess(unit) => unit.typ.value(),
             Self::Structure(structure) => structure.typ.value(),
-        }) * self.get_hp() as u16 / 100
+        }) as usize * self.get_hp() as usize / 100
     }
     pub fn remove_available_mercs(&self, mercs: &mut Vec<MercenaryOption>) {
         for boarded in self.get_boarded() {
